SX5-868 Move the keycloak_utils functions to keycloak.py module_utils.
